refactor(mark): extract evidence query service and update monument entity
- Extract evidence-related query methods from `MarkService` into a new `MarkEvidenceQueryService` interface
- Implement `MarkEvidenceQueryServiceImpl` to encapsulate similarity searches and evidence lookups
- Update `CandidateFetcher`, `ParityChecker`, and `BookmarkContentResolver` to use the new `MarkEvidenceQueryService`
- Remove `MarkEvidenceRepository` dependency and evidence query implementations from `MarkServiceImpl`
- Remove `validationState` field and its related annotations from the `Monument` entity
