+++ monument/src/main/java/pt/estga/monument/entities/Monument.java
@@ -34,9 +34,4 @@ public class Monument extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     private AdministrativeDivision division;
 
-    @Builder.Default
-    @Convert(converter = ValidationStateConverter.class)
-    @Column(nullable = false)
-    private ValidationState validationState = ValidationState.VERIFIED;
-
 }
